refactor: extract ag init into src/commands/init.ts with tmux detection
Extract the init command handler from the monolithic cli.ts into
src/commands/init.ts, adding interactive tmux detection and state
directory creation during bootstrap.

- Move handleInit, template gallery, and prompt logic to commands/init.ts
- Add tmux version detection during ag init (warns if missing/outdated)
- Ensure state directory is created during init
- Wire commands/init.ts into cli.ts dispatch
- bin field already has ag (primary) + aegis (alias)

Closes #2086
